pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = ""
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "passlib[bcrypt]",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx>=0.25.1,<1.0.0",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "sqlmodel<1.0.0,>=0.0.21",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    "pyjwt<3.0.0,>=2.8.0",
    "uvicorn>=0.35.0",
    "python-jose[cryptography]>=3.5.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "celery[redis]>=5.3.6,<6.0.0",
    "redis>=5.0.4",
    "tzdata>=2024.1",
    "sqlalchemy>=2.0.45",
    "factory-boy>=3.3.3",
    "faker>=39.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=1.3.0",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
    "factory-boy>=3.3.0",
    "faker>=25.0.0",
]
[tool.setuptools]
package-dir = { "" = "src" }

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
log_cli = true
filterwarnings = [
    "ignore:'asyncio.get_event_loop_policy'.*:DeprecationWarning:pytest_asyncio.plugin",
]
asyncio_mode = "auto"
